Initial implementation of one-pass streaming analysis, including command line, SpaceSaving, and exponentially decaying emerging itemset mining. Needs significant debugging.

+++ src/main/java/macrobase/analysis/summary/count/ExactCount.java
@@ -6,18 +6,18 @@ import java.util.HashMap;
 import java.util.List;
 
 public class ExactCount {
-    private HashMap<Integer, Integer> counts = new HashMap<>();
+    private HashMap<Integer, Double> counts = new HashMap<>();
 
-    public HashMap<Integer, Integer> getCounts() {
+    public HashMap<Integer, Double> getCounts() {
         return counts;
     }
 
     public ExactCount count(List<DatumWithScore> data) {
         for(DatumWithScore d : data) {
             for(int i : d.getDatum().getAttributes()) {
-                Integer curVal = counts.get(i);
+                Double curVal = counts.get(i);
                 if(curVal == null) {
-                    curVal = 0;
+                    curVal = 0.;
                 }
                 counts.put(i, curVal+1);
             }
